Separate Payment Create and Update

+++ models/PaymentRepository.cs
@@ -3,6 +3,8 @@ using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Cryptography;
+using System.Text;
 
 using static models.Common.PropertyMerger;
 
@@ -10,9 +12,11 @@ namespace models
 {
 	public interface IPaymentRepository
 	{
-		IList<Payment> CreateOrUpdateMany(DateTime time, IList<Payment> updates);
-		IList<Payment> CreateOrUpdateMany(IList<Payment> updates);
+		IList<Payment> CreateMany(DateTime time, IList<Payment> creates);
+		IList<Payment> CreateMany(IList<Payment> creates);
 		IEnumerable<Payment> GetMany(string id);
+		IList<Payment> UpdateMany(DateTime time, IList<Payment> updates);
+		IList<Payment> UpdateMany(IList<Payment> updates);
 	}
 
 	public class PaymentRepository : IPaymentRepository
@@ -26,6 +30,36 @@ namespace models
 			_logger = logger;
 		}
 
+		private static string HashPayment(Payment payment)
+		{
+			var bytes = Encoding.UTF8.GetBytes(
+				$"{payment.Date}-{payment.ExternalId}-{payment.Type}-{payment.SchoolDistrict}");
+			var hash = SHA256.Create().ComputeHash(bytes);
+
+			var sb = new StringBuilder();
+			foreach (var b in hash)
+				sb.Append(b.ToString("x2"));
+
+			return sb.ToString().Substring(0, 10);
+		}
+
+
+		public IList<Payment> CreateMany(DateTime time, IList<Payment> creates)
+		{
+			foreach (var create in creates)
+			{
+				create.PaymentId = HashPayment(create);
+				create.Created = time;
+			}
+
+			_payments.AddRange(creates);
+			return creates;
+		}
+
+		public IList<Payment> CreateMany(IList<Payment> creates) => CreateMany(DateTime.Now, creates);
+
+		public IEnumerable<Payment> GetMany(string id) => _payments.Where(p => p.PaymentId == id).OrderBy(p => p.Split);
+
 		private static IList<string> _excludedFields = new List<string>
 		{
 			nameof(Payment.Id),
@@ -35,8 +69,7 @@ namespace models
 			nameof(Payment.LastUpdated),
 		};
 
-		// TODO(Erik): separate create/update
-		public IList<Payment> CreateOrUpdateMany(DateTime time, IList<Payment> us)
+		public IList<Payment> UpdateMany(DateTime time, IList<Payment> us)
 		{
 			var updated = new List<Payment>();
 
@@ -46,6 +79,9 @@ namespace models
 				var id = kv.Key;
 				var updates = kv.Value;
 				var payments = GetMany(id).ToList();
+				if (payments == null || payments.Count == 0)
+					throw new ArgumentException($"Could not find payments with ID {id}.");
+
 				if (payments.Count > updates.Count)
 					_payments.RemoveRange(payments.Skip(updates.Count));
 
@@ -54,7 +90,7 @@ namespace models
 					var payment = payments.SingleOrDefault(p => p.Split == update.Split);
 					if (payment == null)
 					{
-						// TODO(Erik): create paymentId
+						update.PaymentId = id;
 						update.Created = time;
 						update.LastUpdated = time;
 
@@ -73,8 +109,6 @@ namespace models
 			return updated;
 		}
 
-		public IList<Payment> CreateOrUpdateMany(IList<Payment> updates) => CreateOrUpdateMany(DateTime.Now, updates);
-
-		public IEnumerable<Payment> GetMany(string id) => _payments.Where(p => p.PaymentId == id).OrderBy(p => p.Split);
+		public IList<Payment> UpdateMany(IList<Payment> updates) => UpdateMany(DateTime.Now, updates);
 	}
 }
